v1/requirements.txt
pika==1.3.2
PyYAML==6.0.2
requests==2.32.3
python-dotenv==1.0.1
pymongo==4.8.0